+++ ui/app/components/drop-menu-item.js
@@ -41,8 +41,11 @@ DropMenuItem.prototype.activeNetworkRender = function () {
     case 'Main Ethereum Network':
       if (providerType === 'mainnet') return h('.check', '✓')
       break
-    case 'Morden Test Network':
+    case 'Ropsten Test Network':
       if (provider.type === 'testnet') return h('.check', '✓')
+      break
+    case 'Morden Test Network':
+      if (provider.type === 'morden') return h('.check', '✓')
       break
     case 'Localhost 8545':
       if (activeNetwork === 'http://localhost:8545') return h('.check', '✓')